+++ include/install_functions.sh
@@ -55,6 +55,9 @@ copy_unit_files() {
   if [ "$installation_type" == "BoxTurtle" ]; then
     cp "${afc_path}/templates/AFC_Hardware-AFC.cfg" "${afc_config_dir}/AFC_Hardware.cfg"
     cp "${afc_path}/templates/AFC_Turtle_1.cfg" "${afc_config_dir}/AFC_Turtle_1.cfg"
+  elif [ "$installation_type" == "NightOwl" ]; then
+    cp "${afc_path}/templates/AFC_Hardware-NightOwl.cfg" "${afc_config_dir}/AFC_Hardware.cfg"
+    cp "${afc_path}/templates/AFC_NightOwl_1.cfg" "${afc_config_dir}/AFC_NightOwl_1.cfg"
   fi
 }
 
@@ -90,13 +93,20 @@ install_afc() {
   elif [ "$toolhead_sensor" == "Ramming" ]; then
     update_switch_pin "${afc_config_dir}/AFC_Hardware.cfg" "buffer"
   fi
-  if [ "$buffer_type" == "TurtleNeck" ]; then
+  if [ "$buffer_type" == "TurtleNeck" && "$installation_type" == "BoxTurtle" ]; then
     query_tn_pins "TN"
     append_buffer_config "TurtleNeck" "$tn_advance_pin" "$tn_trailing_pin"
     add_buffer_to_extruder "${afc_config_dir}/AFC_Turtle_1.cfg" "Turtle_1"
-  elif [ "$buffer_type" == "TurtleNeckV2" ]; then
+  elif [ "$buffer_type" == "TurtleNeckV2" && "$installation_type" == "BoxTurtle" ]; then
     append_buffer_config "TurtleNeckV2"
     add_buffer_to_extruder "${afc_config_dir}/AFC_Turtle_1.cfg" "Turtle_1"
+  elif [ "$buffer_type" == "TurtleNeckV2" && "$installation_type" == "NightOwl" ]; then
+    append_buffer_config "TurtleNeckV2"
+    add_buffer_to_extruder "${afc_config_dir}/AFC_NightOwl_1.cfg" "NightOwl_1"
+  elif [ "$buffer_type" == "TurtleNeck" && "$installation_type" == "NightOwl" ]; then
+    query_tn_pins "TN"
+    append_buffer_config "TurtleNeck" "$tn_advance_pin" "$tn_trailing_pin"
+    add_buffer_to_extruder "${afc_config_dir}/AFC_NightOwl_1.cfg" "NightOwl_1"
   fi
   check_and_append_prep "${afc_config_dir}/AFC.cfg"
   replace_varfile_path "${afc_config_dir}/AFC.cfg"
@@ -107,9 +117,18 @@ install_afc() {
 - AFC Configuration updated with selected options at ${afc_file}
 
 - AFC-Klipper-Add-On python extensions installed to ${klipper_dir}/klippy/extras/
+  """
+if [ "$installation_type" == "BoxTurtle" ]; then
+  message+="""
 
 - Ensure you enter either your CAN bus or serial information in the ${afc_config_dir}/AFC_Turtle_1.cfg file
-"""
+  """
+elif [ "$installation_type" == "NightOwl" ]; then
+  message+="""
+
+- Ensure you enter either your CAN bus or serial information in the ${afc_config_dir}/AFC_NightOwl_1.cfg file
+  """
+fi
 if [ "$buffer_type" == "TurtleNeckV2" ]; then
   message+="""
 - Ensure you add the correct serial information to the ${afc_config_dir}/mcu/TurtleNeckv2.cfg file